Reduce logging verbosity
JIRA: APPS-587 Oral History Logging causing p-w-swarmmgr01 to run out of disk space

This commit merely reduces the standard logging in a production
envirobnment. A separate environmental variable will need to be adjust
logging to STDOUT.

The issue of indexing.log is left untouched, and will need to be
resolved.